+++ crates/base/core/src/media/visual/color/palette/mod.rs
@@ -1,0 +1,22 @@
+// devela_base_core::media::visual::color::palette
+//
+//!
+//
+
+// mod ase;
+// mod gpl; // TODO WIP
+// mod hex;
+// mod kpl;
+// mod pal;
+
+crate::structural_mods! { // _mods
+    _mods {
+        // pub use super::{
+        //     // ase::*,
+        //     gpl::*,
+        //     // hex::*,
+        //     // kpl::*,
+        //     // pal::*,
+        // };
+    }
+}